refactor: Extract logic from SupplementController to Action
- Extracts data fetching and usage history calculation logic from `SupplementController` into `App\Actions\Supplements\FetchSupplementsIndexAction`.
- Updates `SupplementController` to use the new Action.
- Removes unused private methods and imports from `SupplementController`.
- Ensures methods adhere to the max 15 lines of code rule.
- Formats code with Pint.
